Refactored into separate pager component
